pyproject.toml
[project]
name = "deep-research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "openai>=1.101.0",
    "openai-agents>=0.2.9",
    "tavily-python>=0.7.11",
]
